% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ins01_parse_from_char.R
\name{ins01_parse_from_char}
\alias{ins01_parse_from_char}
\title{ins01_parse_from_char function to parse
a ins01 tibble with character columns}
\usage{
ins01_parse_from_char(ins01_char)
}
\arguments{
\item{ins01_char}{The ins01 tibble with character columns}
}
\value{
The ins01 tibble with the proper format
}
\description{
ins01_parse_from_char function to parse
a ins01 tibble with character columns
}
