% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ins01_read_from_aws_mysql.R
\name{ins01_read_from_aws_mysql}
\alias{ins01_read_from_aws_mysql}
\title{ins01_read_from_aws_mysql}
\usage{
ins01_read_from_aws_mysql(table)
}
\arguments{
\item{table}{Table name.}
}
\value{
A tibble with the ins01 dataset.
}
\description{
Get the ins01 dataset from a mysql located on A&R MySql cloud server
}
