% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ins01_to_ins02.R
\name{ins01_to_ins02}
\alias{ins01_to_ins02}
\title{ins01_to_ins02}
\usage{
ins01_to_ins02(ins01)
}
\arguments{
\item{ins01}{A tibble with a standard ins01 dataset.}
}
\value{
A tibble with the linear ins02 dataset
}
\description{
Transform the power and prism values to linear components.
SPH/CYL/AX are transformed to the Jackson Cross Cylinder components
SEQ/CP/CC.
PRVM/A is converted to horizontal and vertical components PRVH/V.
}
