% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ins01_to_ins03.R
\name{ins01_to_ins03}
\alias{ins01_to_ins03}
\title{Convert INS01 tibble to INS03 tibble}
\usage{
ins01_to_ins03(ins01)
}
\arguments{
\item{ins01}{Standard ins01 tibble}
}
\value{
ins03 tibble
}
\description{
Convert INS01 tibble to INS03 tibble
}
