% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ins03_range_plot.R
\name{ins03_range_plot}
\alias{ins03_range_plot}
\title{Generate range plot from INS03 dataset}
\usage{
ins03_range_plot(ins03, xmax = Inf)
}
\arguments{
\item{xmax}{Value to be used to filter the ranges to lower values}
}
\value{
a plot
}
\description{
Generate range plot from INS03 dataset
}
